# student/presentation/http/updateStudent.go
package http

import (
	"net/http"
	"strconv"

	"github.com/Abuzar-JS/Go-StudentApp/student/application"
	models "github.com/Abuzar-JS/Go-StudentApp/student/presentation/model"

	"github.com/gin-gonic/gin"
)

// UpdateStudent godoc
// @Summary Update a student
// @Tags students
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param school_id path int true "School ID"
// @Param student_id path int true "Student ID"
// @Param request body models.UpdateStudentRequest true "Student update payload"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string
// @Router /schools/{school_id}/students/{student_id} [put]
func NewUpdateStudent(
	update application.UpdateStudent,
) gin.HandlerFunc {
	return func(ctx *gin.Context) {
		var body models.UpdateStudentRequest
		if err := ctx.ShouldBindJSON(&body); err != nil {
			ctx.JSON(http.StatusBadRequest, gin.H{
				"error": err.Error(),
			})
		}

		studentID := ctx.Param("student_id")
		ID, err := strconv.Atoi(studentID)
		if err != nil {
			ctx.JSON(http.StatusBadRequest, gin.H{
				"message": err.Error(),
			})
			return
		}
		if body.Class == nil && body.Name == nil && body.SchoolID == nil {
			ctx.JSON(http.StatusBadRequest, gin.H{
				"message": "atleast single field is required to update student",
			})
			return
		}

		updateRequest := application.UpdateStudentRequest{
			Name:     body.Name,
			Class:    body.Class,
			SchoolID: body.SchoolID,
		}

		err = update(ctx.Request.Context(), ID, updateRequest)
		if err != nil {
			ctx.JSON(http.StatusBadRequest, gin.H{
				"message": err.Error(),
			})
			return
		}

		ctx.JSON(200, gin.H{
			"message": "Student updated successfully",
		})
	}
}